repos/mvt_data_processing/requirements.txt
aniso8601==9.0.1
argcomplete==3.3.0
blinker==1.8.2
boto3==1.34.126
botocore==1.34.126
cachetools==5.3.3
certifi==2024.6.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
db-dtypes==1.2.0
durationpy==0.6
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.24.0
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
hjson==3.1.0
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.1
pandas==2.2.2
placebo==0.9.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytest==8.3.2
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
troposphere==4.8.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
zappa==0.59.0
